Adding new templates for MPHB custom post types. Review widget setup.
