+++ elements/MPHB/mphb_availability_calendar.php
@@ -38,7 +38,8 @@ class Themo_Widget_MPHB_Availability_Calendar extends Widget_Base {
             'type'        => Controls_Manager::TEXT,
             'label'       => __('Accommodation Type ID', 'th-widget-pack'),
             //'description' => __('ID of Accommodation Type to display availability.', 'th-widget-pack'),
-            'default'     => ''
+            'default'     => '',
+            'label_block' => true,
         ));
 
         $this->add_control('months_to_show', array(
@@ -98,7 +99,21 @@ class Themo_Widget_MPHB_Availability_Calendar extends Widget_Base {
 
     protected function render() {
 
+        global $post;
+
         $settings = $this->get_settings();
+
+        // If Accommodation type id field is empty, try to get the id automatically.
+        if ( !isset( $settings['type_id'] ) || empty( $settings['type_id']) ) {
+            if(isset($post->ID )&& $post->ID > ""){
+                $postID = $post->ID;
+                $themo_post_type = get_post_type($postID);
+                if(isset($themo_post_type) && $themo_post_type=='mphb_room_type'){
+                    $settings['type_id'] = $postID;
+                }
+            }
+        }
+
 
         if ( isset( $settings['type_id'] ) && ! empty( $settings['type_id'] && is_numeric($settings['type_id'])) ) {
